Change FormLabel to Label, change old Label component to ButtonLabel
