+++ src/scss/grommet-core/_objects.button-label.scss
@@ -1,0 +1,24 @@
+// (C) Copyright 2014-2015 Hewlett Packard Enterprise Development LP
+
+.button-label {
+  @include media-query(palm) {
+    &__text {
+      display: none;
+    }
+  }
+
+  @include media-query(lap-and-up) {
+    &__icon {
+      display: none;
+    }
+  }
+
+  a.active > & {
+    .button-label__icon.control-icon {
+      svg {
+        fill: $active-icon-color;
+        stroke: $active-icon-color;
+      }
+    }
+  }
+}
