+++ src/scss/grommet-core/_objects.scss
@@ -9,6 +9,7 @@
 @import "objects.box";
 @import "objects.brick";
 @import "objects.button";
+@import "objects.button-label";
 @import "objects.calendar";
 @import "objects.carousel";
 @import "objects.chart";
@@ -21,7 +22,6 @@
 @import "objects.footer";
 @import "objects.form";
 @import "objects.form-field";
-@import "objects.form-label";
 @import "objects.header";
 @import "objects.headline";
 @import "objects.icon";
